requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.1.0
langchain-community>=0.0.19
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.0
python-multipart>=0.0.5
pydantic>=2.0.0
networkx>=3.0.0
scipy>=1.10.0
numpy>=1.24.0
requests>=2.31.0
